dd installment plan

+++ PoliTechAPI/pt-api/src/main/java/ru/pt/api/dto/payment/InstallmentDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public class InstallmentDto {
     private Long id;
     private Long policyId;
-    private Integer installmentNr;
+    private Long installmentNr;
     private LocalDate dueDate;
     private BigDecimal amount;
     private String currency;
@@ -16,8 +16,8 @@ public class InstallmentDto {
     public void setId(Long id) { this.id = id; }
     public Long getPolicyId() { return policyId; }
     public void setPolicyId(Long policyId) { this.policyId = policyId; }
-    public Integer getInstallmentNr() { return installmentNr; }
-    public void setInstallmentNr(Integer installmentNr) { this.installmentNr = installmentNr; }
+    public Long getInstallmentNr() { return installmentNr; }
+    public void setInstallmentNr(Long installmentNr) { this.installmentNr = installmentNr; }
     public LocalDate getDueDate() { return dueDate; }
     public void setDueDate(LocalDate dueDate) { this.dueDate = dueDate; }
     public BigDecimal getAmount() { return amount; }
