+++ PoliTechAPI/pt-db/src/main/java/ru/pt/db/entity/PaymentInstallmentEntity.java
@@ -15,7 +15,7 @@ public class PaymentInstallmentEntity {
     private Long id;
     private Long tid;
     private Long policyId;
-    private Integer installmentNr;
+    private Long installmentNr;
     private LocalDate dueDate;
     private BigDecimal amount;
     private String currency;
@@ -41,8 +41,8 @@ public class PaymentInstallmentEntity {
     public void setTid(Long tid) { this.tid = tid; }
     public Long getPolicyId() { return policyId; }
     public void setPolicyId(Long policyId) { this.policyId = policyId; }
-    public Integer getInstallmentNr() { return installmentNr; }
-    public void setInstallmentNr(Integer installmentNr) { this.installmentNr = installmentNr; }
+    public Long getInstallmentNr() { return installmentNr; }
+    public void setInstallmentNr(Long installmentNr) { this.installmentNr = installmentNr; }
     public LocalDate getDueDate() { return dueDate; }
     public void setDueDate(LocalDate dueDate) { this.dueDate = dueDate; }
     public BigDecimal getAmount() { return amount; }
